Close to manual setter implementation
