+++ lib/state_inspector/state_inspector.rb
@@ -1,8 +1,50 @@
+require 'read_source'
+
 module StateInspector
   class StateInspector
-    def self.report who, what, old, new
-      key = who.respond_to?(:class_eval) ? who : who.class
-      Reporter[key].update who, what, old, new
+    def initialize base, **opts
+      @base = base # pass in self
+
+      ## setter_filter # Choose whether to enforce setters ending with equals
+      @setter_filter = opts.fetch(:setter_filter) { false }
+    end
+
+    def snoop_setters *setters
+      base.class_eval do
+        setters.
+          select {|m| @setter_filter ? m =~ /=\z/ : m }.
+          select {|m| ![:attr_writer, :attr_accessor].include?(instance_method(m).attr?) }.
+          each do |m|
+            original_method = instance_method(m)
+            define_method(m) do |*args, &block|
+              var = "@#{__method__.to_s.chop}"
+              tell_si var, instance_variable_get(var), *args, &block
+              original_method.bind(self).call(*args, &block)
+            end
+          end
+      end
+    end
+
+    def snoop_methods *meth0ds
+      base.class_eval do
+        meth0ds.
+          each do |m|
+            original_method = instance_method(m)
+            define_method(m) do |*args, &block|
+              tell_si __method__, *args, &block
+              original_method.bind(self).call(*args, &block)
+            end
+          end
+      end
+    end
+
+    private
+    def base
+      if @base.respond_to? :class_eval
+        @base
+      else
+        @base.class
+      end
     end
   end
 end
