+++ test/manual_setters_test.rb
@@ -14,6 +14,7 @@ class ManualSetterTest < Minitest::Test
   def observer; StateInspector::Reporter[M] end
   def test_adds_hook_to_setter_defined_manually
     m = M.new
+    m.state_inspector.snoop_setters :thing=
     m.thing = 42
     assert_empty observer.values
     # add hook
